Use quadrant view for CGAffineTransform interpolation tests
