+++ Example/Unit Tests/CATransform3DInterpolationSnapshotTests.swift
@@ -190,7 +190,7 @@ private final class ContainerView: UIView {
 
     // MARK: - Public Properties
 
-    let animatableView: AnimatableView = .init()
+    let animatableView: QuadrantView = .init()
 
     var transforms: [CATransform3D] {
         get {
@@ -256,55 +256,6 @@ private final class ContainerView: UIView {
         }
 
         gridLayer.path = gridPath.cgPath
-    }
-
-}
-
-// MARK: -
-
-private final class AnimatableView: UIView {
-
-    // MARK: - Life Cycle
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        topLeftLayer.backgroundColor = UIColor.red.cgColor
-        layer.addSublayer(topLeftLayer)
-
-        topRightLayer.backgroundColor = UIColor.green.cgColor
-        layer.addSublayer(topRightLayer)
-
-        bottomLeftLayer.backgroundColor = UIColor.blue.cgColor
-        layer.addSublayer(bottomLeftLayer)
-
-        bottomRightLayer.backgroundColor = UIColor.yellow.cgColor
-        layer.addSublayer(bottomRightLayer)
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Private Methods
-
-    private let topLeftLayer: CALayer = .init()
-    private let topRightLayer: CALayer = .init()
-    private let bottomLeftLayer: CALayer = .init()
-    private let bottomRightLayer: CALayer = .init()
-
-    // MARK: - UIView
-
-    override func layoutSubviews() {
-        [topLeftLayer, topRightLayer, bottomLeftLayer, bottomRightLayer].forEach {
-            $0.bounds.size = .init(width: bounds.width / 2, height: bounds.height / 2)
-        }
-
-        topLeftLayer.position = .init(x: bounds.width * 0.25, y: bounds.height * 0.25)
-        topRightLayer.position = .init(x: bounds.width * 0.75, y: bounds.height * 0.25)
-        bottomLeftLayer.position = .init(x: bounds.width * 0.25, y: bounds.height * 0.75)
-        bottomRightLayer.position = .init(x: bounds.width * 0.75, y: bounds.height * 0.75)
     }
 
 }
